Move isToolAllowedForMode out of shared directory
